requirements.txt
# Core Data Analysis
numpy>=1.21.0
pandas>=2.0.0
polars>=0.19.0

# Machine Learning
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.4.0

# Additional Libraries
requests>=2.26.0
joblib>=1.0.0

# Optional (for extended functionality)
astral>=2.2
python-dotenv>=0.19.0
xarray>=0.19.0
netCDF4>=1.5.7
erddapy>=1.2.0
